Add init-job functionality
Why these changes are being introduced:

The first step in an AB comparison is initializing a 'Job' that
will be used for subsequent 'Runs' of the job.  This requires
setting up a working directory, start a 'job.json' file, and
some supporting utilities.

How this addresses that need:
* Adds first function 'init_job' in abdiff.core
* Adds first utility functions for job names, slug, and
working directory

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-338