Remove job name, prefer explicit job directory
Why these changes are being introduced:

After some internal discussions, it was decided to drop the job name +
slug and instead just provide an explicit job directory to reference a
job.  The net effect is a simplification of functions and utils at this time.

How this addresses that need:
* Reworks core.init_job() to expect a job directory
* Remove mention of job names or slugs anywhere

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-344