Add calc_ab_metrics core function
Why these changes are being introduced:

After a dataset of records and their JSON diffs exists, the next
step is to synthesize aggregate metrics from the diff data.  This
step effectively compresses data for the first time, resulting in
only the need to write metrics JSON to the run JSON file.

How this addresses that need:
* Adds news function calc_ab_metrics
* A sparse matrix of record by fields is created with an integer
value of 1 or 0 indicating if that record had a diff for that field.
This matrix is then queried via DuckDB to generate aggregate metrics.
* After analysis, add new 'metrics' section to run JSON

Side effects of this change:
* Metrics are calculated from a run if invoked

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-351