Add a record links and views to webapp
Why these changes are being introduced:

Once we have diff metrics available, a desired feature of the webapp
is to view any differences for a single record, to understand in more
detail what changed.  Providing a record page is required, but also
some way of generating links that will provide TIMDEX ids to pass
to the record view route.

How this addresses that need:
* Add new Record view route
* Add 'Record Links' section to run page, providing links to
records with diffs for fields or sources
* Add 'Sample' route that expects a field or source and will
generate URLs for Record views that contain a diff for the
field or source.  Given the scale of possible records, this
limits to a max of 100 records per source in the run.

Side effects of this change:
* Record viewing is now possible

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-358