Threads always return container object
Why these changes are being introduced:

Now that we are using threads to run docker containers in
parallel, we still want to return the container object even
if something went wrong so we can collect its logs.

Previously, we were returning a Container object OR an Exception
object from run_docker_container(), which prevented this.

How this addresses that need:
* run_docker_container() now ALWAYS returns a Container object and
MAY return an Exception object
* this supports downstream log aggregation and reporting

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-356