Establish helpers module with TIMDEX source utils
Why these changes are being introduced:

This application is a standalone app to compare A/B versions of the
output of Transmogrifier.  To this end, there are some rote and repetitive
tasks that may need to be performed to provide the inputs this application
needs.

How this addresses that need:
* Establishes a 'abdiff.helper' module that contains code helpful
for working with this application, but not essential to its core functionality
* First sub-module of helpers is code that helps you get lists of ordered
files from TIMDEX S3 bucket

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-379