Handle presence of TXT files from Transmogrifier
Why these changes are being introduced:

It was overlooked that Transmogrifier will write a text file with
records to delete as part of its output, and how this would be
captured in the collating of records and deduping.  In the case of
records where the delete action was the last action, then they
should be removed from the dataset.

How this addresses that need:
* Updates opinionations where a .json extension is assumed
* Updates run_ab_transforms validation to look for output files
that indicate Transmogrifier produced something as output

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-371