Use DeepDiff library for a/b record diffing
Why these changes are being introduced:

Originally, the library jsondiff was used for getting a diff between
the A and B version of a record.  This was selected based on prior
use and experience.  For the most part, it was a nice fit, but lacked
the ability to ignore changes in array order as a diff to report.

This resulted in high numbers of records showing diff, when the only
change was a differently ordered array.

How this addresses that need:
* The DeepDiff library has built-in support to 'ignore_order' when
generating diffs, making it a great fit
* DeepDiff is equally fast for generating diffs
* Our primary, current use of diffs is to highlight records where a
TIMDEX field changed, which is conveniently a "root" field on the
JSON object.  DeepDiff provides an attribute on the diff object
that returns an explicit list of modified "root" fields, which is
just what we need, further simplifying some logic in ABDiff to
parse this information.

Overall, DeepDiff is a more configurable library, with built-in
options that align closely with our use case.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-373