Remove unneeded jsondiff output parsing
Why these changes are being introduced:

Now that the DeepDiff library provides an explicit list of
modified "root" fields -- i.e. TIMDEX fields -- as a built-in
property, we no longer need any additional logic to parse the diff
and surface what fields were modified.

How this addresses that need:
* This removes the helper function generate_field_diff_bools_for_record()
and any tests related to it.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-373