Perform record diff calculations in parallel
Why these changes are being introduced:

Once a collated dataset is prepared, we need to calculate diffs
for each record.  This was memory efficient, but slow, as we
sequentially worked through batches of records.

How this addresses that need:
* By having get_diffed_batches_iter() use parallel processes to
produce batches, the work is sped up considerably, while still
within memory bounds

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-383