Removed transformed files after collation
Why these changes are being introduced:

The size of the raw, original transformed files that are produced by
Transmogrifier are orders of magnitude larger than the parquet datasets
created of the same joined and deduped records.

Unless there is a need to re-use or analyze them, removing them as records
are prepared for collation frees up disk space as the process works.

How this addresses that need:
* Remove transformed files as individual A and B records are written to
the 'transformed' dataset
* Remove the '/transformed' directory entirely once all files are utilized
* Add new env var 'PRESERVE_ARTIFACTS' that, if set to 'true', will skip
this removal.  This optional env var will be used by other artifact removal
steps currently underway.  Default is to remove.

Side effects of this change:
* Reduced storage size of final run

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-366