Add filterable, browsable Records table to Run
Why these changes are being introduced:

One affordance of the webapp is viewing individual records from
a run, seeing a summary of A/B differences, the full A/B records,
and a side-by-side comparison.  But when runs may contain lots of
records, some kind of interface is needed to identify records for
viewing (where once the timdex_record_id is known, the Record page
only requires that).

Previously, a "Record Samples" section appeared on the Run page
that linked to a standalone table of records that met some kind
of criteria (e.g. source = X, or field Y has diffs).  This was
functional, but had drawbacks:

- this static HTML could not handle large number of records, meaning
a representative sample was used, which prevented access to all records
- the combinations of dimensions to drill down into was limited by the
static sample pages of records

How this addresses that need:
* Removes all "Record Samples" approaches
* Replaces with a single table in the Run page
  * filterable by source, whether records had modified specific modified
    fields, even full-text search of the records themselves
* This single table provides a mechanism to browse and filter records
from the run, from a single interface, with arguably simpler logic
under the hood to power it

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-385