Return COALESCE to A/B record joining
Why these changes are being introduced:

When joining the A and B record versions, a previous commit removed a handful
of COALESCE statements in the SQL query, assuming they were redundant given
that A or B record would arbitrarily supply those shared values.  Code review
surfaced that if A or B is missing, then a COALESCE remains helpful to
capture those values.

Side effects of this change:
* Improved A/B record joining

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-383