Allow failed Transmogrifier containers
Why these changes are being introduced:

It was suggested in a previous PR that we raise an exception if ANY
Transmogrifier containers fail.  This appeared to make sense from a data
integrity POV, ensuring that all input files were properly transformed
by Transmogrifier for analysis.  In retrospect, this fell short in two ways:

1. For very large runs, a very small number of containers failing may be
admissible given the run will still contain valuable data and should
be allowed to continue.

2. More subtle, a failed Transmogrifier container may not be an intermittent
bug, but potentially an indication that a code change in Transmog is
problematic.  The responsibility here should be for this application to
surface this in a meaningful way during analysis (out of scope here),
not halt the run completely.

How this addresses that need:
* Creates new env var ALLOW_FAILED_TRANSMOGRIFIER_CONTAINERS, that defaults
to 'true' in the Config class, to allow failed containers
* Continue with run even if failed containers are present

Side effects of this change:
* Runs with failed Transmog containers will continue.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-383