Handle missing A or B record
Why these changes are being introduced:

It is theoretically possible that, after joining the A and B
versions of a record, one of them is absent.  One example might be a code
change in Transmogrifier that fails to transform a particular record.

While uncommon, this application should handle this gracefully and provide
data such that it can be surfaced in the analysis or webapp.

How this addresses that need:
* When calculating the A/B diff, handle if either or both records
are absent.
* Add a new field to the metrics dataset 'a_or_b_missing' that provides
a quick reference boolean for surfacing records where this is the case.
Note: this is currently not utilized in the webapp, but is data now
available.

Side effects of this change:
* Ability to complete large runs with possible missing A or B records.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-388