Update to use parquet dataset output from Transmogrifier
Why these changes are being introduced:

Transmogrifier, and TIMDEX more generally, are getting updated to use
a parquet dataset as the storage architecture for files during ETL.  Transmogrifier
will now start writing to a parquet dataset instead of outputting JSON and TXT
files.  As such, this application needs the ability to read this new Transmog output to
continue providing A/B diff analysis.

Some spike time was spent looking into supporting both "v1" output from Transmog (JSON/TXT files)
and the new parquet dataset writing, but it was determined it's quite a bit of overhead for
functionality that will likely not be used in the coming weeks when we go full v2
parquet dataset writing.

How this addresses that need:

The application has been updated to expect parquet datasets as the output from
each A/B Transmog version.  The changes are mostly localized to the run_ab_transforms()
and collate_ab_transforms() core functions.

In many ways, it was a simplification of code.  A considerable amount of effort was
spent getting the records from v1 JSON/TXT files into a parquet dataset for more
efficient processing.

Side effects of this change:
* Output of JSON and TXT files from v1 Transmogrifier is no longer supported

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-401